Add better documentation
Add more text to the README.md, including installation instructions and
a requirements.txt. I tested a fresh Python env locally and it worked.

As soon as we have the link to the blog post, the placeholder URL should
be replaced.

I also made some small changes to the help text of the CLI.